pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Role-Based Finance Operations Backend"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "python-multipart",
    "python-dotenv",
    "alembic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "bcrypt",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
